Move position class to Board package
