+++ Chess_App/src/main/java/Board/Position.java
@@ -1,4 +1,4 @@
-package Game;
+package Board;
 
 public class Position {
     private short x;
